+++ lib/src/components/live_list/defines.dart
@@ -1,0 +1,14 @@
+// Package imports:
+import 'package:zego_uikit/zego_uikit.dart';
+
+typedef ZegoLiveStreamingOutsideLiveListHost
+    = ZegoOutsideRoomAudioVideoViewStreamUser;
+
+typedef ZegoLiveStreamingOutsideLiveListStyle
+    = ZegoOutsideRoomAudioVideoViewListStyle;
+
+typedef ZegoLiveStreamingOutsideLiveListItemStyle
+    = ZegoOutsideRoomAudioVideoViewListItemStyle;
+
+typedef ZegoLiveStreamingOutsideLiveListConfig
+    = ZegoOutsideRoomAudioVideoViewListConfig;